HH-71: Правки
