+++ app/src/main/java/ru/practicum/android/diploma/features/selectspecialization/data/dto/SpecializationMapDataDomain.kt
@@ -1,7 +1,6 @@
 package ru.practicum.android.diploma.features.selectspecialization.data.dto
 
 import ru.practicum.android.diploma.features.common.data.network.dto.industry.IndustryEntity
-import ru.practicum.android.diploma.features.common.domain.filter.model.FilterIndustry
 import ru.practicum.android.diploma.features.selectspecialization.domain.model.Industry
 
 fun IndustryEntity.toDomain(): Industry {
@@ -11,9 +10,3 @@ fun IndustryEntity.toDomain(): Industry {
     )
 }
 
-fun Industry.toFilterIndustry(): FilterIndustry {
-    return FilterIndustry(
-        id = this.id,
-        name = this.name
-    )
-}
